+++ spec/fixtures/plugins/example_globbing.rb
@@ -1,0 +1,15 @@
+module Danger
+  class Dangerfile
+    module DSL
+      class ExampleGlobbing < Plugin
+        def run
+          return "Hi there globbing 🎉"
+        end
+
+        def self.description
+          "Add a warning to PRs with 'WIP' in their title or body"
+        end
+      end
+    end
+  end
+end